brand for infinite matrices
