+++ Project.toml
@@ -1,19 +1,26 @@
 name = "InfiniteRandomArrays"
 uuid = "2bc77966-89c7-476d-a40f-269028fac4a9"
 authors = ["DanielVandH <danj.vandenheuvel@gmail.com> and contributors"]
-version = "0.1"
+version = "0.2.0"
 
 [deps]
 InfiniteArrays = "4858937d-0d70-526a-a4dd-2d5cb5dd786c"
 LazyArrays = "5078a376-72f3-5289-bfd5-ec5146d43c02"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 
+[weakdeps]
+BandedMatrices = "aae01518-5342-5314-be14-df237901396f"
+
+[extensions]
+InfiniteRandomArraysBandedMatricesExt = "BandedMatrices"
+
 [compat]
-InfiniteArrays = "0.14"
+BandedMatrices = "1.7"
+InfiniteArrays = "0.12, 0.13, 0.14"
 LazyArrays = "2"
 Random = "1"
 Test = "1"
-julia = "1.10"
+julia = "1.6"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
